Extract overlay-source-override image handling from validations.py
Extracts two new modules to trim modules/validations.py:

- modules/validations_overlay_images.py (~700 lines, new):
  overlay-source-override image validation, storage, cleanup, and the
  three Flask endpoint wrappers.  Byte-for-byte extraction of 24
  functions plus the six overlay-image constants
  (MAX_OVERLAY_SOURCE_OVERRIDE_BYTES etc.).

- modules/validations_shared.py (86 lines, new):
  three tiny helpers shared between the overlay-image cluster and
  the YAML file/folder validators still in validations.py --
  _resolve_managed_library_path, _normalize_custom_repo_base,
  _saved_custom_repo_base.  Split out to avoid circular imports
  between validations.py and validations_overlay_images.py.

modules/validations.py: 1828 -> 1247 (-581 / -31.8%).
No behavior change -- validations.py re-exports every moved name so
external callers (blueprints/*, quickstart.py, modules/*, and 35
test-suite monkeypatches) continue to work unchanged.

AST-verified: 24 overlay-image functions are byte-for-byte identical
to develop; the 3 shared helpers are identical modulo added
docstrings/type hints (pure documentation).
